Create new Pyret image type for rendering arbitrary transformations
The code here does not and cannot work as I intended it to. I am committing it just so I can keep it in the record in case I need something from it later. 

This code takes the four points of some image, like a texture to be mapped, as well as four output points, like where a face is to be rendered, and calculates a transformation matrix that takes the input four points and maps them to the output points.

Then it implements a new Pyret image type that I thought would allow for arbitrary matrix transformations. It turns out Canvas2D currently only allows for 2D affine transformations, which means given a square input its output can be at most a parallelogram--it can not produce the distortions needed for it to render arbitrary faces in 3-space. The behavior to allow for this properly seems to be in the works, https://github.com/whatwg/html/issues/5620, but for now I will need to use a hack or some entirely different method for texture mapping.